[refactor] apply global styles across screens.
