+++ src/app/(Authentication)/styles.ts
