+++ src/app/(BottomTabNavigation)/Profile/ResetConfirm/index.tsx
@@ -8,8 +8,12 @@ import BackButton from '../../../../../assets/back-button.svg';
 import Envelope from '../../../../../assets/reset-password-envelope.svg';
 import Refresh from '../../../../../assets/reset-password-refresh.svg';
 import Arrow from '../../../../../assets/right-arrow-white.svg';
-import { ButtonBlack } from '../../../../Components/AuthButton/AuthButton';
+import {
+  ButtonBlack,
+  ButtonTextWhite,
+} from '../../../../Components/AuthButton/AuthButton';
 import { useSession } from '../../../../context/AuthContext';
+import { SafeArea } from '../../../../styles/global';
 
 export default function ResetConfirm() {
   const { email } = useLocalSearchParams() as unknown as { email: string };
@@ -31,46 +35,42 @@ export default function ResetConfirm() {
     setErrorExists(false);
   };
   return (
-    <View style={styles.container}>
+    <SafeArea>
       <View style={styles.contentContainer}>
-        <TouchableOpacity
-          style={styles.backContainer}
-          onPress={() => router.back()}
+        <View>
+          <TouchableOpacity onPress={() => router.back()}>
+            <BackButton />
+          </TouchableOpacity>
+
+          <Text style={styles.titleText}>Reset Password</Text>
+
+          <View style={styles.instructionContainer}>
+            <View style={styles.instructionRow}>
+              <Envelope />
+              <Text style={styles.instructionText}>
+                We will send a six-digit verification code to the email that is
+                registered with your account.
+              </Text>
+            </View>
+
+            <View style={styles.instructionRow}>
+              <Refresh />
+              <Text style={styles.instructionText}>
+                Enter the code in the following screen to create your new
+                password.
+              </Text>
+            </View>
+          </View>
+        </View>
+
+        <ButtonBlack
+          disabled={email === '' || errorExists}
+          onPress={resetPassword}
         >
-          <BackButton />
-        </TouchableOpacity>
-
-        <Text style={styles.titleText}>Reset Password</Text>
-
-        <View style={styles.instructionContainer}>
-          <View style={styles.instructionRow}>
-            <Envelope />
-            <Text style={styles.instructionText}>
-              We will send a six-digit verification code to the email that is
-              registered with your account.
-            </Text>
-          </View>
-
-          <View style={styles.instructionRow}>
-            <Refresh />
-            <Text style={styles.instructionText}>
-              Enter the code in the following screen to create your new
-              password.
-            </Text>
-          </View>
-        </View>
-
-        <View style={styles.nextButton}>
-          <ButtonBlack
-            disabled={email === '' || errorExists}
-            style={styles.nextButtonBase}
-            onPress={resetPassword}
-          >
-            <Text style={styles.nextText}>Continue</Text>
-            <Arrow style={{ marginRight: 12 }} />
-          </ButtonBlack>
-        </View>
+          <ButtonTextWhite>Continue</ButtonTextWhite>
+          <Arrow />
+        </ButtonBlack>
       </View>
-    </View>
+    </SafeArea>
   );
 }
